requirements.txt
python==3.10.16
pytorch==2.0.1
pytorch-cuda==11.8
torchdata==0.6.1
torchtext==0.15.2
matplotlib==3.9.2
spacy==3.7.2
unidecode==1.3.8
seaborn==0.13.2
textstat==0.7.4
notebook==7.3.2
scikit-learn==1.5.2
ipywidgets==8.1.5
torchmetrics==1.6.1
tokenizers==0.20.3
transformers==4.46.3
requests==2.31
pandas==1.4.3
scipy==1.8.1
datasets==2.19.1